+++ backend/src/main/java/com/wd44/drivingschoolsystem/API/Services/StudentService.java
@@ -51,7 +51,10 @@ public class StudentService {
 
     @Transactional
     public @ResponseBody String deleteStudent(int ID) {
-        studentRepo.deleteById(ID);
-        return "Deleted";
+        if (studentRepo.existsById(ID)) {
+            studentRepo.deleteById(ID);
+            return "Deleted";
+        }
+        else return "Not Found";
     }
 }
